LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/events/inline-event-handler-ordering-expected.txt

CONSOLE MESSAGE: line 18: SyntaxError: Unexpected token '}'
CONSOLE MESSAGE: line 51: SyntaxError: Unexpected token '}'

Harness Error (FAIL), message = SyntaxError: Unexpected token '}'

FAIL Inline event handlers retain their ordering when invalid and force-compiled assert_array_equals: property 1, expected "TWO" but got "THREE"
FAIL Inline event handlers retain their ordering when invalid and force-compiled via dispatch assert_array_equals: property 3, expected "TWO" but got "THREE"
FAIL Inline event handlers retain their ordering when invalid and lazy-compiled assert_array_equals: property 1, expected "TWO" but got "THREE"
